cafecompao_api/padarias/api/serializers.py
from rest_framework import serializers
from ..models import Padaria, Cesta, Produto, Categoria, Assinatura, Endereco

class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = '__all__'

class ProdutoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Produto
        fields = '__all__'

class CestaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cesta
        fields = '__all__'

class EnderecoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Endereco
        fields = '__all__'

class PadariaSerializer(serializers.ModelSerializer):
    endereco = EnderecoSerializer()
    class Meta:
        model = Padaria
        fields = '__all__'


class AssinaturaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Assinatura
        fields = '__all__'


